+++ packages/vsx-registry/src/node/vsx-extension-resolver.ts
@@ -14,38 +14,27 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 // *****************************************************************************
 
-import * as os from 'os';
 import * as path from 'path';
 import * as semver from 'semver';
 import * as fs from '@theia/core/shared/fs-extra';
-import { v4 as uuidv4 } from 'uuid';
 import { injectable, inject } from '@theia/core/shared/inversify';
 import URI from '@theia/core/lib/common/uri';
-import { PluginDeployerHandler, PluginDeployerResolver, PluginDeployerResolverContext, PluginDeployOptions } from '@theia/plugin-ext/lib/common/plugin-protocol';
+import { PluginDeployerHandler, PluginDeployerResolver, PluginDeployerResolverContext, PluginDeployOptions, PluginIdentifiers } from '@theia/plugin-ext/lib/common/plugin-protocol';
 import { VSCodeExtensionUri } from '@theia/plugin-ext-vscode/lib/common/plugin-vscode-uri';
 import { OVSXClientProvider } from '../common/ovsx-client-provider';
 import { VSXExtensionRaw } from '@theia/ovsx-client';
 import { RequestService } from '@theia/core/shared/@theia/request';
+import { PluginVSCodeEnvironment } from '@theia/plugin-ext-vscode/lib/common/plugin-vscode-environment';
+import { PluginUninstallationManager } from '@theia/plugin-ext/lib/main/node/plugin-uninstallation-manager';
 
 @injectable()
 export class VSXExtensionResolver implements PluginDeployerResolver {
 
-    @inject(OVSXClientProvider)
-    protected clientProvider: OVSXClientProvider;
-
-    @inject(PluginDeployerHandler)
-    protected pluginDeployerHandler: PluginDeployerHandler;
-
-    @inject(RequestService)
-    protected requestService: RequestService;
-
-    protected readonly downloadPath: string;
-
-    constructor() {
-        this.downloadPath = path.resolve(os.tmpdir(), uuidv4());
-        fs.ensureDirSync(this.downloadPath);
-        fs.emptyDirSync(this.downloadPath);
-    }
+    @inject(OVSXClientProvider) protected clientProvider: OVSXClientProvider;
+    @inject(PluginDeployerHandler) protected pluginDeployerHandler: PluginDeployerHandler;
+    @inject(RequestService) protected requestService: RequestService;
+    @inject(PluginVSCodeEnvironment) protected readonly environment: PluginVSCodeEnvironment;
+    @inject(PluginUninstallationManager) protected readonly uninstallationManager: PluginUninstallationManager;
 
     accept(pluginId: string): boolean {
         return !!VSCodeExtensionUri.toId(new URI(pluginId));
@@ -82,9 +71,9 @@ export class VSXExtensionResolver implements PluginDeployerResolver {
                 return;
             }
         }
-
-        const extensionPath = path.resolve(this.downloadPath, path.basename(downloadUrl));
-        console.log(`[${resolvedId}]: trying to download from "${downloadUrl}"...`, 'to path', this.downloadPath);
+        const downloadPath = (await this.environment.getExtensionsDirUri()).path.fsPath();
+        const extensionPath = path.resolve(downloadPath, path.basename(downloadUrl));
+        console.log(`[${resolvedId}]: trying to download from "${downloadUrl}"...`, 'to path', downloadPath);
         if (!await this.download(downloadUrl, extensionPath)) {
             console.log(`[${resolvedId}]: not found`);
             return;
@@ -94,7 +83,8 @@ export class VSXExtensionResolver implements PluginDeployerResolver {
     }
 
     protected hasSameOrNewerVersion(id: string, extension: VSXExtensionRaw): string | undefined {
-        const existingPlugins = this.pluginDeployerHandler.getDeployedPluginsById(id);
+        const existingPlugins = this.pluginDeployerHandler.getDeployedPluginsById(id)
+            .filter(plugin => !this.uninstallationManager.isUninstalled(PluginIdentifiers.componentsToVersionedId(plugin.metadata.model)));
         const sufficientVersion = existingPlugins.find(existingPlugin => {
             const existingVersion = semver.clean(existingPlugin.metadata.model.version);
             const desiredVersion = semver.clean(extension.version);